test(images): split base-image platform action contract
