+++ test/build-base-image-platform-action.test.ts
@@ -1,0 +1,65 @@
+// SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+import fs from "node:fs";
+import path from "node:path";
+
+import { describe, expect, it } from "vitest";
+import YAML from "yaml";
+
+type Step = {
+  env?: Record<string, unknown>;
+  if?: string;
+  name?: string;
+  run?: string;
+};
+
+const repoRoot = path.resolve(import.meta.dirname, "..");
+
+function required<T>(value: T | undefined, message: string): T {
+  return (
+    value ??
+    (() => {
+      throw new Error(message);
+    })()
+  );
+}
+
+describe("base-image platform action", () => {
+  it("executes dos2unix from each Deep Agents Code platform image before manifest publication (#8870)", () => {
+    const action = YAML.parse(
+      fs.readFileSync(
+        path.join(repoRoot, ".github", "actions", "build-base-image-platform", "action.yaml"),
+        "utf8",
+      ),
+    ) as { runs?: { steps?: Step[] } };
+    const steps = action.runs?.steps ?? [];
+    const validate = required(
+      steps.find((candidate) => candidate.name === "Validate Deep Agents Code dos2unix executable"),
+      "base-image platform action is missing the Deep Agents Code dos2unix validation",
+    );
+    const buildIndex = steps.findIndex(
+      (candidate) => candidate.name === "Build and push platform digest",
+    );
+    const validateIndex = steps.indexOf(validate);
+    const exportIndex = steps.findIndex((candidate) => candidate.name === "Export platform digest");
+
+    expect(validate.if).toBe("${{ inputs.agent == 'langchain-deepagents-code' }}");
+    expect(validate.env).toMatchObject({
+      DIGEST: "${{ steps.build.outputs.digest }}",
+      IMAGE: "${{ inputs.registry }}/${{ inputs.image }}",
+      PLATFORM: "${{ inputs.platform }}",
+    });
+    expect(validate.run).toContain('reference="${IMAGE}@${DIGEST}"');
+    expect(validate.run).toContain('docker run --rm --platform "$PLATFORM"');
+    expect(validate.run).toContain("--network none");
+    expect(validate.run).toContain("--cap-drop ALL");
+    expect(validate.run).toContain("--security-opt no-new-privileges");
+    expect(validate.run).toContain("--read-only");
+    expect(validate.run).toContain("--user 999:999");
+    expect(validate.run).toContain("test -x /usr/bin/dos2unix");
+    expect(validate.run).toContain("dos2unix --version");
+    expect(validateIndex).toBeGreaterThan(buildIndex);
+    expect(validateIndex).toBeLessThan(exportIndex);
+  });
+});
